Moved protocol into fc proper (it's hardly a utility)
